📦 Add plugin 'dressing'
The plugin 'dressing' allows Telescope to act as the input and selector
UI instead of the default Neovim selector/input ui. The telescope UI is
much nicer. This configuration allows for visual selection instead of
index-based selection and works for ANY plugin which uses the
`vim.ui.select()` function for HUD elements -- that's most of them.