✨ Add LSP functionality
This is a whole bear of interdependent plugins. I decided not to use
`lsp-zero` (the all-in-one solution for this) since I don't know how to
configure it to get exactly what I like. I will likely return later to
mess with the aesthetics. For now, just having LSP at all is enough.