📦 Add plugins 'Commentary' and 'Navigator'
Also add bindings for Navigator. Navigator is an improvement over
basic <C-W> commands because it plays nicer with a multiplexer. I
personally use Kitty, so this isn't an issue, but I think I may want
to share this config with other people and there is no guarantee that
they will want to use Kitty.