🖋 Adjust keybinds for 'neotree'
Using `E` and `W` to expand and compress all nodes can cause errors to
pop up. This is a constraint of the OS itself, based on the number of
parallel files that it allows Neovim to handle. The easiest way to
avoid this error is simply to not map that behavior -- oh well.