🖋 Clean keymaps
Neovim's bindings require that you pass it a _function_ to execute
when you hit a button, if you're not using a string. That means that
currying is a great fit here for getting simple functions that can
live in the keybindings. I think it's more readable this way than
putting an anonymous function in each binding, especially since the
anonymous functions take up so much vertical room under stylua's
guidelines (at least 3 lines each, bare minimum).